+++ index.html
@@ -233,10 +233,6 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
-<<<<<<< HEAD
-    <script type="module" src="main.js"></script>
-=======
-  <script src="js/main.js" defer></script>
->>>>>>> 7f6d983 (Задание Больше функций)
+    <script src="js/main.js" type="module"></script>
 </body>
 </html>
